강화 씬 적용, 강화 및 체젠 스크립트 수정

+++ Assets/01KDH/Scripts/GameManager.cs
@@ -23,7 +23,7 @@ public class GameManager : MonoBehaviour
 
     public float _maxhp;
     public float _hp;
-    public int _gold;
+    public float _gold;
 
     public float _goldBonus;
